Args: moved commandline options to own module
